requirements.txt
# Core dependencies
httpx>=0.25.0
python-dotenv>=1.0.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# AI/ML dependencies
google-generativeai>=0.3.0
Pillow>=10.0.0

# Cloud Storage dependencies
google-cloud-storage>=2.10.0

# Development dependencies (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
